+++ Readme/CHANGELOG.md
@@ -52,6 +52,8 @@ If you would like to see fresh changes done since last release you can do so on 
   - `Test-Credential` function to validate credentials
   - `Get-ExportedType.ps1` utility script to list exported types in current session
   - `Get-PropertyType.ps1` utility script to get .NET types of properties
+  - `Out-DataTable` utility function to convert object to data table
+  - `Get-NetworkStatistics.ps1` utility script to run netstat remotely and parse data into object
 
 - Tests
 
@@ -419,7 +421,7 @@ If you would like to see fresh changes done since last release you can do so on 
   - Format-Path function would index into empty array if environment variable not found
   - Minor errors resolved thanks to strict mode
   - Format-Path would return bad result if there is only a single environment variable
-  - Update-Table would produce an error if Get-AllUserSoftware fails to get a list of programs from
+  - Update-Table would produce an error if Get-InstallProperties fails to get a list of programs from
     registry
 
 - Performance